agents/taskweaver/requirements.txt
fastapi
uvicorn
httpx
sse-starlette
python-dotenv
